refector: 반복되는 팀 분포 계산 로직 재사용
